Bsllc development

+++ BOT/_utils/bruhpy.py
@@ -1,6 +1,7 @@
 import re
 import sys
 import time
+import random
 import contextlib
 import python_weather
 import multiprocessing
@@ -19,7 +20,7 @@ def stdoutIO(stdout=None):
     yield stdout
     sys.stdout = old
 
-def execute_processed_command(program, results, debug):
+def execute_processed_command(program, results, debug, pvn):
     """
     Function for executing the code and capturing any output to 
     stdout
@@ -32,15 +33,16 @@ def execute_processed_command(program, results, debug):
         try:
             exec(f"""\n{program}\n""")
             if s.getvalue() != '': results['POST'] = ('NORMAL', '[OUTPUT]\n'+s.getvalue())
-            else:results['POST'] =  ('INFO', '[INFO]: no output produced')
+            else:results[pvn] =  ('INFO', '[INFO]: no output produced')
         except Exception as exception:
             error_response = ''
             line_num = None
             exception = str(exception)
             try:
-                line_num = int(re.search('line (\d)', exception).groups()[0])
+                line_num = int(re.search('line (\d+)\)', exception).groups()[0])
             except Exception as e:
                 pass
+
             error_response += f"-[ERROR]: {exception}\n"
             if "bruhpy" in program:
                 error_response += "it looks like 'bruhpy' was found in the program, did you type it twice?\n"
@@ -49,17 +51,20 @@ def execute_processed_command(program, results, debug):
                     if i == line_num - 2:
                         error_response += f"line {line_num}: '{line}'"
             
-            results['POST'] = ('ERROR', error_response)
+            results[pvn] = ('ERROR', error_response)
+    if debug: print(f'leaving EPC. . .\nwith {pvn} val of {results[pvn]}')
 
 class BruhPy:
     """
     Class responsible for processing of incoming python code from the /bruh command
     """
-    def __init__(self, debug=False):
-        self._debug     = debug
-        self._responses = [] 
-        self._manager   = multiprocessing.Manager()
-        self._results   = self._manager.dict()
+    def __init__(self, debug=False, post_val_name='POST'):
+        self._debug         = debug
+        self._responses     = [] 
+        self._manager       = multiprocessing.Manager()
+        self._results       = self._manager.dict()
+        self._restictions   = ['webbrowser']
+        self._post_val_name = post_val_name
     
     def run(self, arg, argvs):
         """
@@ -67,22 +72,36 @@ class BruhPy:
         :param arg  : second word in the command
         :param argvs: every other word in the command
         """
+        print(arg, argvs)
         if arg == '-s':
-            pre_process = f"{' '.join(argvs) if argvs else ''}".replace('#', '\n').replace('\\t', '\t').replace("“", "\"").replace("”", "\"")
+            pre_process = f"{' '.join(argvs) if argvs else ''}".replace('#', '\n').replace('\\t', '\t').replace("“", "\"").replace("”", "\"").replace("\\\\", "\\")
             self._responses.append(('PY', pre_process))
-        else: pre_process = f"{arg + ' ' + (' '.join(argvs) if argvs else '')}".replace('#', '\n').replace('\\t', '\t').replace("“", "\"").replace("”", "\"")
+        else: pre_process = f"{arg + ' ' + (' '.join(argvs) if argvs else '')}".replace('#', '\n').replace('\\t', '\t').replace("“", "\"").replace("”", "\"").replace("\\\\", "\\")
+
+        code_check = self._check(pre_process)
+
+        if not code_check:
+            self._responses.append(("ERROR", "-[ERROR]: code did not pass preliminary inspection"))
+            self._responses.append(("INFO", "[INFO]: code did not execute, not output produced"))
+            return self._responses
 
         # Execute the code 
-        process = multiprocessing.Process(target=execute_processed_command, args=(pre_process, self._results, self._debug))
+        process = multiprocessing.Process(target=execute_processed_command, args=(pre_process, self._results, self._debug, self._post_val_name))
         process.start()
 
         # sleep while code is running
-        time.sleep(2)
+        time.sleep(1)
 
         # timeout after the two seconds
         if process.is_alive():
             process.terminate()
             self._responses.append(('ERROR', '-[ERROR]: valid runtime exceeded!'))
-        else: self._responses.append(self._results['POST'])
+        else: self._responses.append(self._results[self._post_val_name])
 
         return self._responses
+
+    def _check(self, program):
+        for restriction in self._restictions:
+            if re.search(restriction, program):
+                return False
+        return True
